pyproject.toml
[tool.poetry]
name = "resumeanalyser"
version = "0.1.0"
description = "The ResumeAnalyser package is a comprehensive toolkit designed for automated and efficient parsing, analysis, and evaluation of resumes. Leveraging advanced algorithms, it extracts key information like educational background, professional experience, skills, and certifications from a variety of resume formats. This package also includes features for ranking candidates based on customizable criteria, making it an indispensable tool for HR professionals and recruiters seeking to streamline the hiring process. Its intuitive API and compatibility with multiple programming languages ensure easy integration into existing HR systems, significantly enhancing recruitment workflows."
authors = ["group14"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11.6"
python-docx = "^1.1.0"
pypdf = "^3.17.4"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
myst-nb = {version = "^1.0.0", python = "^3.11.6"}
sphinx-autoapi = "^3.0.0"
sphinx-rtd-theme = "^2.0.0"
pytest = "^7.4.4"

[tool.semantic_release]
version_variable = "pyproject.toml:version" # version location
branch = "main"                             # branch to make releases of
changelog_file = "CHANGELOG.md"             # changelog file
build_command = "poetry build"              # build dists
dist_path = "dist/"                         # where to put dists
upload_to_release = true                    # auto-create GitHub release
upload_to_pypi = false                      # don't auto-upload to PyPI
remove_dist = false                         # don't remove dists
patch_without_tag = true                    # patch release by default

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
